pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "genai-assessment"
version = "0.1.0"
description = "Azure Document Intelligence + Azure OpenAI based field extraction and medical chatbot."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "GenAI Assessment" }]
dependencies = [
  "azure-ai-documentintelligence>=1.0.2,<2.0.0",
  "azure-core>=1.30.0",
  "openai>=1.70.0,<3.0.0",
  "pydantic>=2.11,<3.0",
  "pydantic-settings>=2.6,<3.0",
  "tenacity>=9.0,<10.0",
  "diskcache>=5.6,<6.0",
  "streamlit>=1.40,<2.0",
  "structlog>=24.1,<26.0",
  "python-dotenv>=1.0,<2.0",
  "httpx>=0.27,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9.0",
  "pytest-asyncio>=0.24,<2.0",
  "pytest-cov>=5.0,<7.0",
  "ruff>=0.7,<1.0",
  "mypy>=1.11,<2.0",
  "types-requests",
  "pre-commit>=3.7,<5.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["common*", "form_extraction*", "medical_chatbot*"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["common", "form_extraction", "medical_chatbot"]

[tool.ruff.lint]
select = [
  "E", "F", "W",
  "I",
  "B",
  "UP",
  "SIM",
  "ASYNC",
  "RUF",
  "N",
]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "N802", "RUF001"]
# The assignment mandates camelCase JSON field names; mirroring them in the
# Pydantic schema is intentional.
"form_extraction/backend/schemas.py" = ["N815", "RUF001"]
"form_extraction/frontend/**" = ["RUF001"]
"medical_chatbot/**" = ["RUF001"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_return_any = true
disallow_untyped_defs = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]
exclude = ["build/", "dist/"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra --strict-markers"
markers = [
  "integration: requires RUN_AZURE_TESTS=1 and live Azure credentials",
]
